seperated functions into different files to make code neater
