+++ midterm.py
@@ -3,13 +3,11 @@ import os
 import random
 import re
 import string
-import warnings
 import webbrowser
 
 import numpy as np
 import pandas as pd
 import plotly.express as px
-import statsmodels.api
 from pandas.core.dtypes.common import (
     is_bool_dtype,
     is_categorical_dtype,
@@ -17,11 +15,18 @@ from pandas.core.dtypes.common import (
     is_numeric_dtype,
     is_string_dtype,
 )
-from plotly import figure_factory as ff
 from plotly import graph_objects as go
 from scipy import stats
 from sklearn.ensemble import RandomForestClassifier
-from sklearn.preprocessing import LabelEncoder
+
+from cat_cat import cat_cat_dwm, cat_correlation, heat_mx
+from cont_cat import (
+    cat_cont_corr_ratio,
+    cont_cat_dwm,
+    cont_cat_graph,
+    logistic_regression,
+)
+from cont_cont import cont_cont_dwm, linear_regression
 
 
 def process_dataframe(
@@ -376,230 +381,6 @@ def process_dataframe(
     return
 
 
-# make categorical vs categorical graph
-def heat_mx(df, predictor, response):
-    crosstab_df = pd.crosstab(
-        index=df[predictor], columns=df[response], margins=False
-    )  # noqa: E501
-
-    fig = go.Figure(
-        data=go.Heatmap(
-            x=crosstab_df.index,
-            y=crosstab_df.columns,
-            z=crosstab_df.values,
-            zmin=0,
-            zmax=crosstab_df.max().max(),
-        )
-    )  # noqa: E501
-
-    fig.update_layout(
-        title=f"{predictor} {response} Heatmap",
-        xaxis_title=f"{predictor}",
-        yaxis_title=f"{response}",
-    )
-
-    filename = f"plots/{predictor}_{response}_heatmap.html"
-
-    fig.write_html(
-        file=filename,
-        include_plotlyjs="cdn",
-    )
-
-    return filename
-
-
-# linear regression model
-def linear_regression(df, predictor, response):
-    predictor_for_model = statsmodels.api.add_constant(df[predictor])
-    linear_regression_model = statsmodels.api.OLS(
-        df[response], predictor_for_model
-    )  # noqa: E501
-    linear_regression_model_fitted = linear_regression_model.fit()
-
-    # Get the stats
-    t_value = round(linear_regression_model_fitted.tvalues[1], 6)
-    p_value = "{:.6e}".format(linear_regression_model_fitted.pvalues[1])
-
-    filename = f"plots/{predictor}_{response}_linear_regression.html"
-
-    # Plot the figure
-    fig = px.scatter(
-        data_frame=df,
-        x=predictor,
-        y=response,
-        trendline="ols",
-        title=f"(t-value={t_value}) (p-value={p_value})",
-    )
-
-    fig.write_html(
-        file=filename,
-        include_plotlyjs="cdn",
-    )
-
-    return filename
-
-
-def cont_cat_graph(df, predictor, response):
-    filename1 = f"plots/{predictor}_{response}_dist_plot.html"
-    filename2 = f"plots/{predictor}_{response}_violin_plot.html"
-    group_labels = df[response].unique()
-    hist_data = []
-
-    for label in group_labels:
-        sub_df = df[df[response] == label]
-        group = sub_df[predictor]
-        hist_data.append(group)
-
-    # Create distribution plot
-    fig_1 = px.histogram(df, x=predictor, color=response, marginal="rug")
-    fig_1.update_layout(
-        title=f"{predictor} by {response}",
-        xaxis_title=predictor,
-        yaxis_title=response,
-    )
-
-    fig_1.write_html(
-        file=filename1,
-        include_plotlyjs="cdn",
-    )
-
-    fig_2 = px.violin(df, y=predictor, color=response, box=True)
-
-    fig_2.update_layout(
-        title=f"{response} by {predictor} ",
-        xaxis_title=predictor,
-        yaxis_title=response,
-    )
-
-    fig_2.write_html(
-        file=filename2,
-        include_plotlyjs="cdn",
-    )
-
-    return filename1, filename2
-
-
-# Used correlation function from Lecture 7
-def fill_na(data):
-    if isinstance(data, pd.Series):
-        return data.fillna(0)
-    else:
-        return np.array([value if value is not None else 0 for value in data])
-
-
-def cat_correlation(x, y, bias_correction=True, tschuprow=False):
-    """
-    Calculates correlation statistic for categorical-categorical association.
-    The two measures supported are:
-    1. Cramer'V ( default )
-    2. Tschuprow'T
-
-    SOURCES:
-    1.) CODE: https://github.com/MavericksDS/pycorr
-    2.) Used logic from:
-        https://stackoverflow.com/questions/20892799/using-pandas-calculate-cram%C3%A9rs-coefficient-matrix
-        to ignore yates correction factor on 2x2
-    3.) Haven't validated Tschuprow
-
-    Bias correction and formula's taken from :
-    https://www.researchgate.net/publication/270277061_A_bias-correction_for_Cramer's_V_and_Tschuprow's_T
-
-    Wikipedia for Cramer's V: https://en.wikipedia.org/wiki/Cram%C3%A9r%27s_V
-    Wikipedia for Tschuprow' T: https://en.wikipedia.org/wiki/Tschuprow%27s_T
-    Parameters:
-    -----------
-    x : list / ndarray / Pandas Series
-        A sequence of categorical measurements
-    y : list / NumPy ndarray / Pandas Series
-        A sequence of categorical measurements
-    bias_correction : Boolean, default = True
-    tschuprow : Boolean, default = False
-               For choosing Tschuprow as measure
-    Returns:
-    --------
-    float in the range of [0,1]
-    """
-    corr_coeff = np.nan
-    try:
-        # using pandas since only passing df through this program
-        if pd.isna(x).any():
-            x = fill_na(x)
-        if pd.isna(y).any():
-            y = fill_na(y)
-        crosstab_matrix = pd.crosstab(x, y)
-        n_obs = crosstab_matrix.sum().sum()
-
-        yates_correct = True
-        if bias_correction:
-            if crosstab_matrix.shape == (2, 2):
-                yates_correct = False
-
-        chi2, _, _, _ = stats.chi2_contingency(
-            crosstab_matrix, correction=yates_correct
-        )
-        phi2 = chi2 / n_obs
-
-        # r and c are number of categories of x and y
-        r, c = crosstab_matrix.shape
-        if bias_correction:
-            phi2_correct = max(0, phi2 - ((r - 1) * (c - 1)) / (n_obs - 1))
-            r_correct = r - ((r - 1) ** 2) / (n_obs - 1)
-            c_correct = c - ((c - 1) ** 2) / (n_obs - 1)
-            if tschuprow:
-                corr_coeff = np.sqrt(
-                    phi2_correct / np.sqrt((r_correct - 1) * (c_correct - 1))
-                )
-                return corr_coeff
-            corr_coeff = np.sqrt(
-                phi2_correct / min((r_correct - 1), (c_correct - 1))
-            )  # noqa: E501
-            return corr_coeff
-        if tschuprow:
-            corr_coeff = np.sqrt(phi2 / np.sqrt((r - 1) * (c - 1)))
-            return corr_coeff
-        corr_coeff = np.sqrt(phi2 / min((r - 1), (c - 1)))
-        return corr_coeff
-    except Exception as ex:
-        print(ex)
-        if tschuprow:
-            warnings.warn("Error calculating Tschuprow's T", RuntimeWarning)
-        else:
-            warnings.warn("Error calculating Cramer's V", RuntimeWarning)
-        return corr_coeff
-
-
-def cat_cont_corr_ratio(categories, values):
-    """
-    Correlation Ratio: https://en.wikipedia.org/wiki/Correlation_ratio
-    SOURCE:
-    1.) https://towardsdatascience.com/
-    the-search-for-categorical-correlation-a1cf7f1888c9
-    :param categories: Numpy array of categories
-    :param values: Numpy array of values
-    :return: correlation
-    """
-    f_cat, _ = pd.factorize(categories)
-    cat_num = np.max(f_cat) + 1
-    y_avg_array = np.zeros(cat_num)
-    n_array = np.zeros(cat_num)
-    for i in range(0, cat_num):
-        cat_measures = values[np.argwhere(f_cat == i).flatten()]
-        n_array[i] = len(cat_measures)
-        y_avg_array[i] = np.average(cat_measures)
-    y_total_avg = np.sum(np.multiply(y_avg_array, n_array)) / np.sum(n_array)
-    numerator = np.sum(
-        np.multiply(
-            n_array, np.power(np.subtract(y_avg_array, y_total_avg), 2)
-        )  # noqa: E501
-    )
-    denominator = np.sum(np.power(np.subtract(values, y_total_avg), 2))
-    if numerator == 0:
-        eta = 0.0
-    else:
-        eta = np.sqrt(numerator / denominator)
-    return eta
-
-
 def make_heatmap_html(matrix: pd.DataFrame):
     fig = go.Figure(
         data=go.Heatmap(
@@ -613,328 +394,6 @@ def make_heatmap_html(matrix: pd.DataFrame):
     )
     matrix_html = fig.to_html()
     return matrix_html
-
-
-def cont_cont_dwm(
-    df: pd.DataFrame, pred1: str, pred2: str, response: str
-):  # noqa: E501
-    pred1_edges = np.histogram_bin_edges(df[pred1], bins=10)
-    pred2_edges = np.histogram_bin_edges(df[pred2], bins=10)
-
-    bin_mean, pred1_edges, pred2_edges, b = stats.binned_statistic_2d(
-        df[pred1],
-        df[pred2],
-        df[response],
-        statistic="mean",
-        bins=(pred1_edges, pred2_edges),
-    )
-
-    bin_count, pred1_edges, pred2_edges, b = stats.binned_statistic_2d(
-        df[pred1],
-        df[pred2],
-        df[response],
-        statistic="count",
-        bins=(pred1_edges, pred2_edges),
-    )
-
-    pop_mean = df[response].mean()
-
-    # using later for graphs and also for calculations
-    pred1_centers = (pred1_edges[:-1] + pred1_edges[1:]) / 2
-    pred2_centers = (pred2_edges[:-1] + pred2_edges[1:]) / 2
-    pred1_centers = pred1_centers.tolist()
-    pred2_centers = pred2_centers.tolist()
-
-    # calculating diff w/ mean (not weighted & weighted)
-    diff = bin_mean - pop_mean
-    sq_diff = np.square(diff)
-    sum_sq_diff = np.nansum(sq_diff)
-    total_bin = bin_mean.size
-    msd = sum_sq_diff / total_bin
-    pop_prop = bin_count / len(df[response])
-    w_sq_diff = pop_prop * sq_diff
-    wmsd = np.nansum(w_sq_diff)
-
-    pop_prop = pop_prop.round(3)
-    # note: plotly defines x opposite of binned stats so
-    # I have swapped pred 1 & 2 so now pred2 is x ....
-    fig = ff.create_annotated_heatmap(
-        x=pred2_centers,
-        y=pred1_centers,
-        z=bin_mean,
-        annotation_text=bin_mean.round(3),
-        colorscale="curl",
-        showscale=True,
-        customdata=pop_prop,
-        hovertemplate="<b>x</b>: %{x}<br>"
-        + "<b>y</b>: %{y}<br>"
-        + "<b>z</b>: %{z}"
-        + "<extra>Population<br>Proportion: %{customdata}</extra>",
-        hoverongaps=False,
-        zmid=pop_mean,
-        zmin=df[response].min(),
-        zmax=df[response].max(),
-    )
-
-    fig.update_layout(
-        title_text=f"{pred1} & {pred2} Bin Averages of Response",
-        xaxis=dict(tickmode="array", tickvals=np.around(pred2_edges, 3)),
-        yaxis=dict(tickmode="array", tickvals=np.around(pred1_edges, 3)),
-    )
-    filename1 = f"plots/{pred1}_{pred2}_diff_of_mean_resp_bin.html"
-
-    fig.write_html(
-        file=filename1,
-        include_plotlyjs="cdn",
-    )
-
-    # residual plot
-    fig_2 = ff.create_annotated_heatmap(
-        x=pred2_centers,
-        y=pred1_centers,
-        z=diff,
-        colorscale="curl",
-        customdata=pop_prop,
-        hovertemplate="<b>x</b>: %{x}<br>"
-        + "<b>y</b>: %{y}<br>"
-        + "<b>z</b>: %{z}"
-        + "<extra>Population<br>Proportion: %{customdata}</extra>",
-        annotation_text=diff.round(3),
-        hoverongaps=False,
-        showscale=True,
-        zmid=pop_mean,
-        zmin=df[response].min(),
-        zmax=df[response].max(),
-    )
-
-    fig_2.update_layout(
-        title_text=f"{pred1} & {pred2} Bin Average Residual",
-        xaxis=dict(tickmode="array", tickvals=np.around(pred2_edges, 3)),
-        yaxis=dict(tickmode="array", tickvals=np.around(pred1_edges, 3)),
-    )
-    filename2 = f"plots/{pred1}_{pred2}_dwm_of_resp_residual.html"
-
-    fig_2.write_html(
-        file=filename2,
-        include_plotlyjs="cdn",
-    )
-
-    return filename1, filename2, round(msd, 3), round(wmsd, 3)
-
-
-# predictor 1 is continous predictor 2 is categorical
-def cont_cat_dwm(df: pd.DataFrame, pred1: str, pred2: str, response: str):
-    pred1_edges = np.histogram_bin_edges(df[pred1], bins="sturges")
-    # getting values for each category
-    categories = df[pred2].unique()
-    label_encoder = LabelEncoder()
-    int_encoded = label_encoder.fit_transform(categories)
-
-    # using later for graphs and also for calculations
-    pred1_centers = (pred1_edges[:-1] + pred1_edges[1:]) / 2
-    pred1_centers = pred1_centers.tolist()
-    all_bin = pd.DataFrame(index=int_encoded, columns=pred1_centers)
-    bin_count = pd.DataFrame(index=int_encoded, columns=pred1_centers)
-    bin_mean = pd.DataFrame(index=int_encoded, columns=pred1_centers)
-
-    for x in range(len(categories)):
-        for i in range(len(pred1_edges) - 1):
-            temp_bin = []
-            temp_bin = df[
-                (df[pred2] == categories[x])
-                & (df[pred1] >= pred1_edges[i])
-                & (df[pred1] < pred1_edges[i + 1])
-            ][response]
-            all_bin.at[int_encoded[x], pred1_centers[i]] = temp_bin
-            bin_count.loc[int_encoded[x], pred1_centers[i]] = len(temp_bin)
-            bin_mean.loc[int_encoded[x], pred1_centers[i]] = np.mean(temp_bin)
-
-    pop_mean = df[response].mean()
-
-    # calculating diff w/ mean (not weighted & weighted)
-    diff = bin_mean - pop_mean
-    sq_diff = np.square(diff)
-    sum_sq_diff = np.nansum(sq_diff)
-    total_bin = len(pred1_centers) * len(categories)
-    msd = sum_sq_diff / total_bin
-    pop_prop = bin_count / len(df[response])
-    w_sq_diff = pop_prop * sq_diff
-    wmsd = np.nansum(w_sq_diff)
-
-    pop_prop = pop_prop.astype("float").round(3)
-    round_bin_mean = bin_mean.astype("float").round(3)
-    fig = ff.create_annotated_heatmap(
-        x=pred1_centers,
-        y=categories.tolist(),
-        z=bin_mean.values.tolist(),
-        annotation_text=round_bin_mean.values.tolist(),
-        colorscale="curl",
-        showscale=True,
-        customdata=pop_prop,
-        hovertemplate="<b>x</b>: %{x}<br>"
-        + "<b>y</b>: %{y}<br>"
-        + "<b>z</b>: %{z}"
-        + "<extra>Population<br>Proportion: %{customdata}</extra>",
-        hoverongaps=False,
-        zmid=pop_mean,
-        zmin=df[response].min(),
-        zmax=df[response].max(),
-    )
-
-    fig.update_layout(
-        title_text=f"{pred1} & {pred2} Bin Averages of Response",
-        xaxis=dict(tickmode="array", tickvals=np.around(pred1_edges, 3)),
-    )
-    filename1 = f"plots/{pred1}_{pred2}_diff_of_mean_resp_bin.html"
-
-    fig.write_html(
-        file=filename1,
-        include_plotlyjs="cdn",
-    )
-
-    # residual plot
-    round_diff = diff.astype("float").round(3)
-    fig_2 = ff.create_annotated_heatmap(
-        x=pred1_centers,
-        y=categories.tolist(),
-        z=diff.values.tolist(),
-        colorscale="curl",
-        customdata=pop_prop,
-        hovertemplate="<b>x</b>: %{x}<br>"
-        + "<b>y</b>: %{y}<br>"
-        + "<b>z</b>: %{z}"
-        + "<extra>Population<br>Proportion: %{customdata}</extra>",
-        annotation_text=round_diff.values.tolist(),
-        hoverongaps=False,
-        showscale=True,
-        zmid=pop_mean,
-        zmin=df[response].min(),
-        zmax=df[response].max(),
-    )
-
-    fig_2.update_layout(
-        title_text=f"{pred1} & {pred2} Bin Average",
-        xaxis=dict(tickmode="array", tickvals=np.around(pred1_edges, 3)),
-    )
-    filename2 = f"plots/{pred1}_{pred2}_dwm_of_resp_residual.html"
-
-    fig_2.write_html(
-        file=filename2,
-        include_plotlyjs="cdn",
-    )
-
-    return filename1, filename2, round(msd, 3), round(wmsd, 3)
-
-
-def cat_cat_dwm(df: pd.DataFrame, pred1: str, pred2: str, response: str):
-    # getting unique values for each category
-    categories1 = df[pred1].unique().astype(str)
-    categories2 = df[pred2].unique().astype(str)
-    categories1 = sorted(categories1)
-    categories2 = sorted(categories2)
-
-    # get mean and bin count
-    bin_mean = pd.crosstab(
-        index=df[pred1],
-        columns=df[pred2],
-        values=df[response],
-        aggfunc="mean",
-        margins=False,
-    )
-    pop_mean = df[response].mean()
-    bin_count = pd.crosstab(index=df[pred1], columns=df[pred2], margins=False)
-
-    # calculating diff w/ mean (not weighted & weighted)
-    diff = bin_mean - pop_mean
-    sq_diff = np.square(diff)
-    sum_sq_diff = np.nansum(sq_diff)
-    total_bin = len(categories2) * len(categories1)
-    msd = sum_sq_diff / total_bin
-    pop_prop = bin_count / len(df[response])
-    w_sq_diff = pop_prop * sq_diff
-    wmsd = np.nansum(w_sq_diff)
-
-    pop_prop = pop_prop.round(3)
-    fig = ff.create_annotated_heatmap(
-        x=bin_mean.columns.tolist(),
-        y=bin_mean.index.tolist(),
-        z=bin_mean.values,
-        annotation_text=bin_mean.values.round(3),
-        colorscale="curl",
-        showscale=True,
-        customdata=pop_prop.values,
-        hovertemplate="<b>x</b>: %{x}<br>"
-        + "<b>y</b>: %{y}<br>"
-        + "<b>z</b>: %{z}"
-        + "<extra>Population<br>Proportion: %{customdata}</extra>",
-        hoverongaps=False,
-        zmid=pop_mean,
-        zmin=df[response].min(),
-        zmax=df[response].max(),
-    )
-
-    fig.update_layout(title_text=f"{pred1} & {pred2} Bin Averages of Response")
-    filename1 = f"plots/{pred1}_{pred2}_diff_of_mean_resp_bin.html"
-
-    fig.write_html(
-        file=filename1,
-        include_plotlyjs="cdn",
-    )
-
-    # residual plot
-    # diff_w_pop = str(diff) + "(pop: " + str(bin_count) + ")"
-    fig_2 = ff.create_annotated_heatmap(
-        x=categories2,
-        y=categories1,
-        z=diff.values,
-        colorscale="curl",
-        customdata=pop_prop,
-        hovertemplate="<b>x</b>: %{x}<br>"
-        + "<b>y</b>: %{y}<br>"
-        + "<b>z</b>: %{z}"
-        + "<extra>Population<br>Proportion: %{customdata}</extra>",
-        annotation_text=diff.values.round(3),
-        hoverongaps=False,
-        showscale=True,
-        zmid=pop_mean,
-        zmin=df[response].min(),
-        zmax=df[response].max(),
-    )
-
-    fig_2.update_layout(title_text=f"{pred1} & {pred2} Bin Average")
-    filename2 = f"plots/{pred1}_{pred2}_dwm_of_resp_residual.html"
-
-    fig_2.write_html(
-        file=filename2,
-        include_plotlyjs="cdn",
-    )
-
-    return filename1, filename2, round(msd, 3), round(wmsd, 3)
-
-
-def make_html_link(plot_col: pd.Series):
-    # regex for making link text
-    regex = ".+/([^/]+).html$"
-    for x in range(len(plot_col)):
-        text = re.findall(regex, plot_col[x])
-        link_html = (
-            f'<a target="_blank" href="{plot_col[x]}">{text[0]}</a>'  # noqa: E501
-        )
-        plot_col[x] = link_html
-    return plot_col
-
-
-def logistic_regression(pandas_df, predictor, response):
-    predictor_for_model = statsmodels.api.add_constant(pandas_df[predictor])
-    logistic_regression_model = statsmodels.api.Logit(
-        pandas_df[response], predictor_for_model
-    )
-    logistic_regression_model_fitted = logistic_regression_model.fit()
-
-    # Get the stats
-    t_value = round(logistic_regression_model_fitted.tvalues[1], 6)
-    p_value = "{:.6e}".format(logistic_regression_model_fitted.pvalues[1])
-    return p_value, t_value
 
 
 def cont_dwm(pandas_df, predictor, response):
@@ -1026,3 +485,15 @@ def rf_importance(df, predictors, response):
     clf.fit(df_X, df_y)
     importances = clf.feature_importances_
     return importances.tolist()
+
+
+def make_html_link(plot_col: pd.Series):
+    # regex for making link text
+    regex = ".+/([^/]+).html$"
+    for x in range(len(plot_col)):
+        text = re.findall(regex, plot_col[x])
+        link_html = (
+            f'<a target="_blank" href="{plot_col[x]}">{text[0]}</a>'  # noqa: E501
+        )
+        plot_col[x] = link_html
+    return plot_col
